v0.0.892 — ArtNet Monitor/Console, bug fixes, warning cleanup
Phase 3 #6 — ArtNet Monitor (nodeType 18) + ArtNet Console (nodeType 19):
- ArtNetMonitorNode.h/.cpp: pass-through + ArtNetMonitorBuffer (dirty-flag
  atomic ring); universe filter (-1=all, or specific universe); "--" display
  when filter active but no matching universe received
- ArtNetConsoleNode.h/.cpp: 512 atomic fader channels + blackout + universe
  atomic; outputs PAX_TYPE_DMX with key=universe; #includes ArtNetMonitorNode.h
- New React: ArtNetMonitorNode.tsx, ArtNetConsoleNode.tsx (reuse DmxShared.tsx)
- Bridge.ts: ArtNetSnapshotEntry, onArtNetSnapshot, setArtNetConsoleChannel,
  setArtNetBlackout
- App.tsx: nodeTypes 18/19, artNetTimers flash, rawToFlowNode cases
- Sidebar.tsx + GenericNode.tsx: ArtNet Monitor/Console entries
- WebBridge: pushArtNetSnapshots, artNetConsoleChannel/artNetBlackout handlers,
  handleUndo/handleRedo extended to nodeType 19, addNode defaults for 18/19
- PatchyProcessor: artNetMonitorBuffers/artNetConsoleBuffers maps,
  drainAllArtNetSnapshots, saveArtNetConsoleChannels,
  restoreArtNetConsoleChannels (with graph ptr overload)
- PatchyEditor: all ArtNet console callbacks wired
- GraphModel: cases 18/19 ports + labels
- CMakeLists.txt: ArtNetMonitorNode.cpp, ArtNetConsoleNode.cpp added
- ArtNetDeviceNodes.cpp: auto-configure universe 0 on fresh node drop

Undo/redo (same pattern as DMX Console):
- artNetChannels base64 in settingsJson; transfer/restore/reset callbacks
- transferLastSent/transferBlackout across graph rebuilds
- setNodeParam(artNetConsoleChannel/artNetBlackout) excluded from pushSnapshot
- fullSettingsRef includes live blackout state (fixes phantom undo step)

Bug fixes:
- Fold button broken on DMX/ArtNet nodes: pass _forceCollapsed to useNodeCollapsed
- BO resets on node drop: restoreBlackout called in restoreChannels path
- Spurious flash on node drop with BO active (multi-layer fix):
  restoreChannels compares zeros vs zeros when BO active;
  transferBlackout zeros lastSent; read blackout from settingsJson not node atomic
- Phantom undo step on BO toggle: dmxBlackout/artNetBlackout added to
  setNodeParam exclusion guard (was hitting pushSnapshot despite early return)

Warning cleanup (34 warnings → 0):
- static_cast<size_t> on all array indexing in DmxConsoleNode.h,
  ArtNetConsoleNode.h, PatchyProcessor.h
- (tcflag_t) casts in SerialPort.h
- unsigned for dmxLen in ArtNetDeviceNodes.h
- Renamed v→elem in WebBridge.cpp, n→nIn in AudioDeviceNodes.cpp
- Added case PortType::Midi + default→"unknown" in GraphModel.cpp
- Removed empty for loop (unused n) in GraphModel.cpp
- ~WebBridge() override in WebBridge.h

Tested: ArtNet Console → QLC+ ✓  QLC+ → ArtNet Monitor ✓
        Full loop Console→Out→QLC+→DMX→Lyre ✓
        Undo/redo DMX+ArtNet Console ✓  Blackout ✓  Fold ✓
        0 compiler warnings ✓

Docs: Architecture.md + README updated for v0.0.892

+++ Source/ArtNetDeviceNodes.h
@@ -77,11 +77,11 @@ namespace ArtNetCodec
                           uint8_t* buf, int bufCap)
     {
         // DMX data comes from v.data[]; pad to even length, minimum 2
-        int dmxLen = v.dataSize > 0 ? (int) v.dataSize : 1;
-        if (dmxLen & 1) ++dmxLen;   // must be even
-        dmxLen = std::min (dmxLen, 512);
+        unsigned dmxLen = v.dataSize > 0 ? (unsigned) v.dataSize : 1u;
+        if (dmxLen & 1u) ++dmxLen;   // must be even
+        dmxLen = std::min (dmxLen, 512u);
 
-        if (bufCap < kHdrSize + dmxLen) return 0;
+        if ((unsigned) bufCap < (unsigned) kHdrSize + dmxLen) return 0;
 
         // ID "Art-Net\0"
         std::memcpy (buf, kArtNetId, 8);
@@ -107,13 +107,13 @@ namespace ArtNetCodec
         buf[17] = (uint8_t) (dmxLen & 0xFFu);
 
         // DMX data — copy what we have, zero-pad the rest
-        int srcLen = std::min ((int) v.dataSize, dmxLen);
+        unsigned srcLen = std::min ((unsigned) v.dataSize, dmxLen);
         if (srcLen > 0)
             std::memcpy (buf + kHdrSize, v.data, (size_t) srcLen);
         if (srcLen < dmxLen)
             std::memset (buf + kHdrSize + srcLen, 0, (size_t) (dmxLen - srcLen));
 
-        return kHdrSize + dmxLen;
+        return (int) (kHdrSize + dmxLen);
     }
 
 } // namespace ArtNetCodec
@@ -186,7 +186,7 @@ public:
             if (universe >= 0 && (int) pkt.value.key != universe)
                 continue;
 
-            outputValues[count] = pkt.value;
+            outputValues[static_cast<size_t>(count)] = pkt.value;
             ++count;
         }
         outputValueCount = count;
